fix: rename `PostFilters` to `EventFilters` and create new inteface for post filters
